#16 rename core -> src
